requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncclick==8.1.7.2
attrs==23.2.0
cffi==1.16.0
cryptography==43.0.0
frozenlist==1.4.1
googleapis-common-protos==1.63.2
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
multidict==6.0.5
protobuf==5.27.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyserial==3.5
python-dotenv==1.0.1
python-kasa==0.7.0.5
sniffio==1.3.1
typing_extensions==4.12.2
viam-sdk==0.25.2
yarl==1.9.4
